Cambio de estructura, comentarios y megustas

+++ server.js
@@ -1,15 +1,13 @@
 const express = require('express');
 const fs = require('fs');
 const multer = require('multer');
-const path = require('path'); 
-var port = process.env.port || 3000;
-var cookieParser = require('cookie-parser');
-var session = require('express-session');
+const path = require('path');
+const cookieParser = require('cookie-parser');
+const session = require('express-session');
 const app = express();
-var server = require('http').Server(app);
-const { conectarDB, Usuario, Image } = require('./mongo');
-//const storage = multer.memoryStorage();
-//const upload = multer({ storage: storage });
+const server = require('http').Server(app);
+const { conectarDB, Usuario } = require('./mongo');
+const port = process.env.port || 3000;
 
 //CONFIGURACIONES
 app.use(express.json());
@@ -19,8 +17,8 @@ app.use(session({
     secret: 'clave_mateo_sergio', 
     resave: false,
     saveUninitialized: true
-    /*cookie: { secure: false } // Cambia a true si usas HTTPS en producción*/
 }));
+
 var auth = function(req, res, next){
     if (req.session && req.session.user) {
         return next();
@@ -28,13 +26,13 @@ var auth = function(req, res, next){
         return res.sendStatus(401); // No autorizado
     }
 }
-// Configuración imágenes
+
+// Configuración de Multer para imágenes
 const storage = multer.diskStorage({
     destination: function (req, file, cb) {
-        
-        const username = req.session.user; 
-        const userFolderPath = `public/images/${username}`; // Carpeta específica para el usuario
-        fs.mkdir(userFolderPath, { recursive: true }, (err) => { // Crea la carpeta del usuario si no existe
+        const username = req.session.user;
+        const userFolderPath = `public/images/${username}`;
+        fs.mkdir(userFolderPath, { recursive: true }, (err) => {
             if (err) {
                 console.error("Error al crear la carpeta del usuario:", err);
             }
@@ -42,13 +40,11 @@ const storage = multer.diskStorage({
         });
     },
     filename: function (req, file, cb) {
-        // Nombre de archivo personalizado (puedes usar cualquier lógica aquí)
         cb(null, Date.now() + path.extname(file.originalname));
     }
 });
 
 const upload = multer({ storage: storage });
-
 
 //FUNCIONES GET
 
@@ -57,12 +53,13 @@ app.get('/', (req, res) => {
     res.setHeader('Content-type', 'text/html');
     res.send(contenido);
 });
-app.get("/registro", (req,response)=> {
+
+app.get("/registro", (req, response) => {
     var contenido = fs.readFileSync("public/registro.html");
     response.setHeader("Content-type", "text/html");
     response.send(contenido);
 });
-// Ruta para servir la página del muro
+
 app.get('/feed', auth, (req, res) => {
     var contenido = fs.readFileSync('public/feed.html', 'utf8');
     res.setHeader('Content-type', 'text/html');
@@ -70,37 +67,33 @@ app.get('/feed', auth, (req, res) => {
 });
 
 app.get('/home', auth, (req, res) => {
-    const username = req.session.user; // Obtiene el nombre de usuario de la sesión
+    const username = req.session.user;
     if (username) {
         var contenido = fs.readFileSync('public/home.html', 'utf8');
-        // Insertar un script al final del body que defina una variable global con el nombre de usuario
         contenido = contenido.replace('</body>', `<script>var currentUser = "${username}";</script></body>`);
         res.setHeader('Content-type', 'text/html');
         res.send(contenido);
     } else {
-        res.sendStatus(401); // Si no hay usuario en la sesión, responde con un estado no autorizado
+        res.sendStatus(401);
     }
 });
 
-app.get('/imagenes-usuario', async (req, res) => {
+app.get('/imagenes-usuario', auth, async (req, res) => {
     try {
-        const username = req.session.user; // Obtén el nombre de usuario de la sesión
-        // Busca el usuario en la base de datos por su nombre de usuario
+        const username = req.session.user;
         const usuario = await Usuario.findOne({ username });
         if (usuario) {
-            // Si se encuentra el usuario, devuelve las rutas relativas de las imágenes
-            const imagenesRelativas = usuario.publicaciones.map(ruta => ruta.replace('public\\', ''));
+            const imagenesRelativas = usuario.publicaciones.map(publicacion => publicacion.imagePath.replace('public\\', ''));
             res.json({ imagenes: imagenesRelativas });
         } else {
-            // Si no se encuentra el usuario, devuelve un mensaje de error
             res.status(404).json({ error: 'Usuario no encontrado' });
         }
     } catch (error) {
-        // Maneja cualquier error
         console.error('Error al obtener las imágenes del usuario:', error);
         res.status(500).json({ error: 'Error del servidor' });
     }
 });
+
 app.get('/cargarFeed', auth, async (req, res) => {
     try {
         const usuarios = await Usuario.find({}, 'publicaciones username').exec();
@@ -108,9 +101,7 @@ app.get('/cargarFeed', auth, async (req, res) => {
 
         usuarios.forEach(usuario => {
             usuario.publicaciones.forEach(publicacion => {
-                const rutaRelativa = publicacion.replace(/^public[\\/]/, '');
-                const imagePath = `images/${usuario.username}/${path.basename(publicacion)}`;
-                console.log('Ruta de la imagen:', imagePath);
+                const imagePath = publicacion.imagePath.replace(/^public[\\/]/, '');
                 publicaciones.push({ 
                     username: usuario.username, 
                     imagePath: imagePath 
@@ -127,10 +118,9 @@ app.get('/cargarFeed', auth, async (req, res) => {
 
 //FUNCIONES POST
 
-//Registrar en Bdd
 app.post("/registrar", async function(req, res){
     console.log("Datos recibidos para registro: ", req.body);
-    if(!req.body.username || !req.body.password || !req.body.nombre || !req.body.apellidos || !req.body.fechaNacimiento || !req.body.etiquetas){
+    if (!req.body.username || !req.body.password || !req.body.nombre || !req.body.apellidos || !req.body.fechaNacimiento || !req.body.etiquetas) {
         res.send({"res":"register failed"});
     } else {
         try {
@@ -158,59 +148,62 @@ app.post("/registrar", async function(req, res){
     }
 });
 
-
-//Identificación
 app.post("/identificar", async function(req, res){
-    if(!req.body.username || !req.body.password){//Si un campo no esta rellenado esto evita problemas en el servidor
-            res.send({"res":"login failed"});
-    }else {
-        const usuarioEncontrado=await Usuario.findOne({username:req.body.username}, {password:req.body.password});
-            if(usuarioEncontrado){
-                req.session.user = req.body.username;
-                req.session.userId = usuarioEncontrado._id; // Almacenamos el ID del usuario en la sesión
-                req.session.admin = true;
-                return res.send({"res":"login true"});
-            }else{
-        res.send({"res":"usuario no válido"});
-            }
+    if (!req.body.username || !req.body.password) {
+        res.send({"res":"login failed"});
+    } else {
+        const usuarioEncontrado = await Usuario.findOne({ username: req.body.username, password: req.body.password });
+        if (usuarioEncontrado) {
+            req.session.user = req.body.username;
+            req.session.userId = usuarioEncontrado._id;
+            req.session.admin = true;
+            return res.send({"res":"login true"});
+        } else {
+            res.send({"res":"usuario no válido"});
+        }
     }
 });
-//Gestión imágenes
+
 app.post('/upload', auth, upload.single('imagen'), async (req, res) => {
     try {
-        // Obtenemos la ruta de la imagen cargada
         const imagePath = req.file.path;
-
-        // Obtenemos el ID del usuario que ha iniciado sesión
         const usuarioId = req.session.userId;
-
-        // Actualizamos la base de datos con la URL de la imagen
-        await Usuario.findByIdAndUpdate(usuarioId, { $push: { publicaciones: imagePath } });
-
+        
+        const nuevaPublicacion = {
+            imagePath: imagePath
+        };
+        
+        await Usuario.findByIdAndUpdate(usuarioId, { $push: { publicaciones: nuevaPublicacion } });
+        
         res.status(200).json({ message: 'Imagen subida correctamente', imagePath: imagePath });
     } catch (error) {
         console.error(error);
         res.status(500).json({ message: 'Error al subir la imagen' });
     }
 });
+
 app.delete('/delete-photo/:photoId', auth, async (req, res) => {
     try {
         const usuarioId = req.session.userId;
         const photoId = req.params.photoId;
-
-        // Elimina la foto de la base de datos
-        await Usuario.findByIdAndUpdate(usuarioId, { $pull: { publicaciones: photoId } });
-
-        // Obtiene la ruta de la foto a eliminar
-        const photoPath = path.join(__dirname, 'public', 'images', photoId);
-
-        // Verifica si el archivo existe antes de intentar eliminarlo
-        if (fs.existsSync(photoPath)) {
-            // Elimina la foto físicamente del sistema de archivos
-            fs.unlinkSync(photoPath);
+        
+        const usuario = await Usuario.findById(usuarioId);
+        const publicacion = usuario.publicaciones.id(photoId);
+        
+        if (publicacion) {
+            const photoPath = publicacion.imagePath;
+            
+            usuario.publicaciones.id(photoId).remove();
+            await usuario.save();
+            
+            if (fs.existsSync(photoPath)) {
+                fs.unlinkSync(photoPath);
+            }
+            
+            res.status(200).json({ message: 'Foto eliminada correctamente' });
+        } else {
+            res.status(404).json({ message: 'Foto no encontrada' });
         }
-
-        res.status(200).json({ message: 'Foto eliminada correctamente' });
     } catch (error) {
         console.error(error);
         res.status(500).json({ message: 'Error al eliminar la foto' });
